semi final version of aprovetime
