+++ Server.js
@@ -44,6 +44,15 @@ import {
   ApproveTimeSheet
 } from "./database/databaseTimeSheet.js";
 
+import { 
+  IsAdmin,
+  IsManager,
+  isAuthenticated, 
+} from './routes/Authentication.js';
+
+
+
+
 import { CreatePDF } from "./pdf/pdfTest.js";
 import { ConvertJsonToExcel } from "./xlsx/xlsxTest.js";
 import path from "node:path";
@@ -56,6 +65,12 @@ const __dirname = fileURLToPath(new URL(".", import.meta.url));
 import session from "express-session";
 import { stringify } from "querystring";
 import { Console, log } from "console";
+
+import managerRequests from './routes/ManagerRequests.js';
+import adminRequests from './routes/AdminRequests.js';
+
+
+
 //import { autoMailer } from './e-mail_notification/mail.js';
 
 // The server is given the name app and calls from the express function
@@ -190,117 +205,18 @@ app.get("/profileData", async (req, res) => {
 });
 
 
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
 // handle the manager function
 
 app.use("/manager",IsManager, serveStatic(join(__dirname, "manager")));
-
-//Maneger skal kunne se brugere under sig og hvilke projekter der er under sig
-
-app.post("/managerRequests", isAuthenticated, async (req, res) => {
-  console.log(req.body);
-
-
-switch (req.body.functionName) {
-  case "LinkUsers":
-    let ProjectManagerID1 = await GetUserIdWithName(poolData, req.body.managerToLink);
-    let userID1 = await GetUserIdWithName(poolData, req.body.userToLink);
-    let projectID = await GetProjectIdWithName(
-      poolData,
-      req.body.projectToLink
-    );
-    let newLinkData = await CreateUserProjectManagerlink(
-      poolData,
-      userID1,
-      ProjectManagerID1,
-      projectID
-    );
-    console.log(newLinkData);
-    break;
-
-
-  case "ApproveTimeSheet":
-   
-      let TimeSheetApprove = await ApproveTimeSheet(poolData,req.body.TimeSheetId);
-      
-      console.log(TimeSheetApprove);    
-
-  break;
-  default:
-    break;
-}
-
-});
-
-app.get("/managerRequests",IsManager, async (req, res)=>{
-   console.log(req.query);
-  switch (req.query.functionName) {
-
-    case "GetProjectManagerProjects":
-      let ProjectManagerID2 = await GetUserIdWithName(poolData, req.session.userName);
-      let managerProjects = await GetManagerProjects(poolData, ProjectManagerID2);
-      console.log(managerProjects);
-      res.send(managerProjects);
-      break; 
-
-    case "GetUsersUnderManager":
-      let ManagerID3 = await GetUserIdWithName(poolData, req.session.userName);
-      let Users = await GetUsersUnderManager(poolData,ManagerID3);
-      console.log(Users);
-      res.send(Users);
-    break;
-
-    case "GetUserInfo":
-
-    let usernames = [];
-    let users = req.query.users.split(","); // split the string by comma
-    console.log(users[0]+users.length); // should log 82
-    for (let i = 0; i < users.length; i++) {
-      usernames[i] = await GetUsernameWithID(poolData,users[i]);
-    }
-    res.send(usernames)
-    break;
-    
-    case "GetTimeSheet":
-    let TimeSheetData = await GetFilledOutTimeSheetForUser(poolData,req.query.UserID,moment().isoWeek(),moment().year());
-    res.send(TimeSheetData);
-    break;
-
-    case "GetProjectInfo":
-
-    console.log(req.query.TaskId);
-
-    // Get Task name and project name
-    let TasknameAndProjectName = await GetTaskNameAndProjectName(poolData,req.query.TaskId);
-
-    console.log(TasknameAndProjectName);
-
-    res.send(TasknameAndProjectName);
+app.use("", managerRequests);
 
 
 
-    break;
-  
-    default:
-      break;
-  }
-})
+
+
+
+
+
 
 
 
@@ -325,158 +241,11 @@ app.get("/managerRequests",IsManager, async (req, res)=>{
 
 // This folder is only accelisble after the user is confirmed to be an admin
 app.use("/admin", IsAdmin, serveStatic(join(__dirname, "admin")));
+app.use("", adminRequests);
+
 
 // Handle the Admins requsts
-app.post("/adminRequests", isAuthenticated, async (req, res) => {
-  switch (req.body.functionName) {
-    case "CreateUser":
-      let CreateUserData = await CreateUser(
-        poolData,
-        req.body.createUsername,
-        req.body.createPassword,
-        0,
-        req.body.FullName,
-        req.body.PhoneNumber,
-        req.body.Email
-      );
-      console.log(CreateUserData);
-      res.status(201).send("User: " + req.body.createUsername + " has been created");
-      break;
-    case "CreateProject":
 
-        console.log(req.body);
-
-      let ProjectManagerID5 = await GetUserIdWithName(poolData,req.body.projectmanager);
-
-      let CreateProjectData = await CreateProject(
-        poolData,
-        req.body.projectName,
-        req.body.projectStartDate,
-        req.body.projectEndDate,
-        req.body.projectHoursSpent,
-        ProjectManagerID5,
-      );
-      console.log(CreateProjectData);
-      res.status(201).send("Project: " + req.body.projectName + " has been created");
-      break;
-
-
-    case "seeUserLevel":
-      let userID1 = await GetUserIdWithName(poolData, req.body.seeUserLevel);
-      if (userID1 === false) {
-        res.status(400).send("User: "+ req.body.seeUserLevel + " does not exist");
-        break;
-      }
-      let seeUserLevelData = await GetUserLevel(poolData, userID1);
-      if (seeUserLevelData === false) {
-        res.status(500).send("Internal Server Error");
-        break;
-      }
-      res.send(seeUserLevelData);
-      break;
-
-
-
-    case "setUserLevel":
-      let userID2 = await GetUserIdWithName(
-        poolData,
-        req.body.setUserLevelName
-      );
-      console.log(userID2);
-      let seeUserNewLevelData = await SetUserLevel(poolData,userID2,  req.body.setUserIsAdmin,req.body.SetUserIsManager);
-      console.log(seeUserNewLevelData);
-      let check1 = req.body.setUserIsAdmin; let check2 = req.body.SetUserIsManager;
-      res.status(201).send("User: " + req.body.setUserLevelName + " Is Now " + (check1 ? "Admin, " : "") + (check2 ? "Manager, " : ""));
-      break;
-      
-      case "LinkUserToManagerForm":
-
-      let ManagerID = await GetUserIdWithName(poolData, req.body.Manager);
-
-      if (ManagerID == false) {
-        res.status(400).send("User: "+ req.body.Manager + " does not exist");
-        break;
-      }
-      
-      let UserID5 = await GetUserIdWithName(poolData, req.body.User);
-
-      if (UserID5 == false) {
-        res.status(400).send("User: "+ req.body.User + " does not exist");
-        break;
-      }
-
-
-
-      console.log(ManagerID);
-
-      let usermanagerlink = await CreateUserManagerlink(poolData, UserID5, ManagerID );
-      
-      if (usermanagerlink == true) {
-        res.status(201).send("User: " + req.body.User + " is now under manager: " + req.body.Manager);
-      }else{
-        res.status(500).send("Error has occured and changes have not been made")
-      }
-
-      break;
-
-
-
-
-    case "ExportPDF":
-      let userID3 = await GetUserIdWithName(
-        poolData,
-        req.session.userName
-      );
-      GetUserProjects(poolData, userID3).then((projects) => {
-        console.log(projects);
-        CreatePDF(req.session.userName, projects).then((pdfPath) => {
-          const stream = fs.createReadStream(pdfPath);
-          stream.on("open", () => {
-            stream.pipe(res);
-          });
-          stream.on("error", (err) => {
-            res.end(err);
-          });
-          res.on("finish", () => {
-            fs.unlink(pdfPath, (err) => {
-              if (err) throw err;
-              console.log("PDF file deleted");
-            });
-          });
-        });
-      });
-
-      break;
-    case "ExportExcel":
-      let userID4 = req.session.userName;
-      GetProjects(poolData).then((projects) => {
-        JSON.stringify(projects);
-        ConvertJsonToExcel(projects, userID4).then((xlsxPath) => {
-          console.log(xlsxPath);
-          res.download(xlsxPath);
-        });
-      });
-      break;
-    case "CreateTasks":
-      let projectID2 = await GetProjectIdWithName(
-        poolData,
-        req.body.projectToLink
-      );
-      let task = await CreateTasks(
-        poolData,
-        projectID2,
-        req.body.taskName,
-        req.body.taskDescription,
-        req.body.estimate
-      );
-      console.log(task);
-      res.status(201).send("Task: " + req.body.taskName + " Has now been created for " + req.body.projectToLink);
-    default:
-      break;
-  }
-
-  console.log(req.body);
-});
 
 
 
@@ -556,29 +325,3 @@ app.use((req, res) => {
   res.status(404).send("404 error page does not exist");
 });
 
-function isAuthenticated(req, res, next) {
-  if (req.session.isAuthenticated) {
-    next();
-  } else {
-    // send error message
-    res.status(401).send("Acces not granted");
-  }
-}
-
-function IsAdmin(req, res, next) {
-  if (req.session.UserLevel.isAdmin) {
-    next();
-  } else {
-    // send error message
-    res.status(401).send("Acces not granted");
-  }
-}
-
-function IsManager(req, res, next) {
-  if (req.session.UserLevel.isManager) {
-    next();
-  } else {
-    // send error message
-    res.status(401).send("Acces not granted");
-  }
-}
