+++ manager/Aprovetime.html
@@ -12,9 +12,8 @@
 
 
     <div id="UserHolder"></div>
-    <table id="CurrentTimeSheet">        
+    <table id="TimeSheetContainer">        
     </table>    
-    <div id="ButtonHolder"> </div>
 
     <script src="Aprovetime.js"></script>
 </body>
